grapheditor: moved more state into global state in preparation for fully controlled graph editor with viewport control
